Rename aas_teaclave -> worker_teaclave
